backend/requirements.txt
google-generativeai==0.8.3
qdrant-client==1.11.0
markdown==3.7
python-dotenv==1.0.1
pydantic==2.9.2
fastapi==0.115.0
uvicorn==0.30.6